moving around files (hit file length limit) fixed up confirm disapear when participant admitted
